+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Since v1.6.1
 
+- Internal Representation change:
+  - Added grid to BoundaryState with type Grid and module Tgraph.Grid 
+    (for much faster touching vertex checks).
+
+- Other changes
+
   - Added joinDashing', dashj', dashJ', defaultJoinDashSize
   - Added completeEdgeSet, bothDirSet
   - Exported tryOnBoundary, isBoundaryDE, isBoundaryV
@@ -12,7 +18,7 @@
 - Internal Representation change:
     - Changed type of BoundaryDedges in BoundaryState (constructors not exported)
        (This does not affect boundary or boundaryESet functions).
-    - boundaryAt is much more efficient (using new representation for BoundaryDedges) and returns the two boundary directed edges in direction order.
+    - boundaryAt is much more efficient (using new representation for BoundaryDedges) and returns the two boundary directed edges in boundary direction order.
 
 - Other changes:
     - Added nextBV and prevBV (efficient)